+++ pulled_search.py
@@ -13,7 +13,8 @@
 
     Usage:
         pulled_search.py -c file -d path
-            {-P [-m path] [-z] [-a] | -I [-n path]}
+            {-P [-m path] [-a] [-z] |
+             -I [-n path]}
             [-t email {email2 email3 ...} {-s subject_line}]
             [-y flavor_id]
             [-v | -h]
@@ -21,26 +22,26 @@
     Arguments:
         -c file => Configuration file.  Required argument.
         -d dir_path => Directory path for option '-c'.  Required argument.
+
         -P => Process Doc ID files send to RabbitMQ.
-        -m dir_path => Directory to monitor for doc ID files.
-            Used for the -P option.
-        -a => This is an archive log search.
-            Used for the -P option.
-        -z => Use the zgrep option instead of check_log to check GZipped files.
-            Used for the -P option.
+            -m dir_path => Directory to monitor for doc ID files.
+            -a => This is an archive log search.
+            -z => Use the zgrep option instead of check_log to check GZipped
+                files.
+
         -I => Insert Pulled Search files into Mongodb.
-        -n dir_path => Directory to monitor for pulled search files.
-            Used for the -I option.
+            -n dir_path => Directory to monitor for pulled search files.
+
         -t email_address(es) => Send output to one or more email addresses.
-        -s subject_line => Pre-amble to the subject line of email.
-            Requires -t option.
+            -s subject_line => Pre-amble to the subject line of email.
+
         -y value => A flavor id for the program lock.  To create unique lock.
         -v => Display version of this program.
         -h => Help and usage message.
 
         NOTE 1:  -v or -h overrides the other options.
         NOTE 2:  -s requires -t option to be included.
-        NOTE 3:  -P and -I are Xor options.
+        NOTE 3:  -P and -I are XOR options.
         NOTE 4:  -m and -n options will override the configuration settings.
             The -m option is mapped to the doc_dir configuration entry, and
             the -n option is mapped to the monitor_dir configuration entry.
@@ -82,7 +83,7 @@
             # Pulled Search Process/RabbitMQ Configuration section.
             # Update this section if using the -P option.
             user = "USER"
-            pswd = "PSWD"
+            japd = "PSWORD"
             host = "HOSTNAME"
             # RabbitMQ Queue name.
             queue = "QUEUENAME"
@@ -120,39 +121,75 @@
             #   config file.
             mconfig = "mongo"
 
-        Configuration file (config/mongo.py.TEMPLATE).  Below is the
-        configuration file format for the Mongo instance setup.
-            Update this file if using the -I option.
+        Mongo configuration file format (config/mongo.py.TEMPLATE).  The
+            configuration file format is for connecting to a Mongo database or
+            replica set for monitoring.  A second configuration file can also
+            be used to connect to a Mongo database or replica set to insert the
+            results of the performance monitoring into.
 
-            # Pulled Search Insert/Mongo DB Configuration section.
-            user = "USERNAME"
-            passwd = "PASSWORD"
-            # Mongo DB host information
+            There are two ways to connect methods:  single Mongo database or a
+            Mongo replica set.
+
+            Single database connection:
+
+            # Single Configuration file for Mongo Database Server.
+            user = "USER"
+            japd = "PSWORD"
             host = "HOST_IP"
             name = "HOSTNAME"
-            # Mongo database port (default is 27017)
             port = 27017
-            # Mongo configuration settings
             conf_file = None
-            # Authentication required:  True|False
             auth = True
+            auth_db = "admin"
+            auth_mech = "SCRAM-SHA-1"
+            use_arg = True
+            use_uri = False
 
-            # Replica Set Mongo configuration settings.
-            # Replica set name.
-            #    None means the Mongo database is not part of a replica set.
-            #    Example:  repset = "REPLICA_SET_NAME"
-            repset = None
-            # Replica host listing.
-            #    None means the Mongo database is not part of a replica set.
-            #    Example:  repset_hosts = "HOST1:PORT, HOST2:PORT, [...]"
-            repset_hosts = None
-            # Database to authentication to.
-            #    Example:  db_auth = "AUTHENTICATION_DATABASE"
-            db_auth = None
+            Replica set connection:  Same format as above, but with these
+                additional entries at the end of the configuration file.  By
+                default all these entries are set to None to represent not
+                connecting to a replica set.
+
+            repset = "REPLICA_SET_NAME"
+            repset_hosts = "HOST1:PORT, HOST2:PORT, HOST3:PORT, [...]"
+            db_auth = "AUTHENTICATION_DATABASE"
+
+            Note:  If using SSL connections then set one or more of the
+                following entries.  This will automatically enable SSL
+                connections. Below are the configuration settings for SSL
+                connections.  See configuration file for details on each entry:
+
+            ssl_client_ca = None
+            ssl_client_key = None
+            ssl_client_cert = None
+            ssl_client_phrase = None
+
+            Note:  FIPS Environment for Mongo.
+              If operating in a FIPS 104-2 environment, this package will
+              require at least a minimum of pymongo==3.8.0 or better.  It will
+              also require a manual change to the auth.py module in the pymongo
+              package.  See below for changes to auth.py.
+
+            - Locate the auth.py file python installed packages on the system
+                in the pymongo package directory.
+            - Edit the file and locate the "_password_digest" function.
+            - In the "_password_digest" function there is an line that should
+                match: "md5hash = hashlib.md5()".  Change it to
+                "md5hash = hashlib.md5(usedforsecurity=False)".
+            - Lastly, it will require the Mongo configuration file entry
+                auth_mech to be set to: SCRAM-SHA-1 or SCRAM-SHA-256.
+
+            # Name of Mongo database for data insertion
+            db = "DATABASE"
+            # Name of Mongo table/collection.
+            tbl = "COLLECTION"
+
+        Configuration modules -> Name is runtime dependent as it can be used to
+            connect to different databases with different names.
 
     Examples:
         pulled_search.py -c search -d /opt/local/pulled/config -P
-            -t Mark.J.Pernot@coe.ic.gov -s Pulled Search Notification
+            -t myname@email.domain -s Pulled Search Notification
 
         pulled_search.py -c search -d /opt/local/pulled/config -I
             -n /opt/local/pulled/monitor -y pulled_insert