+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 ## [0.1.3] - 2021-12-15
+### Changed
+- config/mongo.py.TEMPLATE:  Added SSL connection entries.
 
 
 ## [0.1.2] - 2020-07-01