+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 
 ## [0.1.3] - 2021-12-15
 ### Changed
+- process_insert:  Captured and processed status return from mongo_libs.ins_doc call.
 - config/mongo.py.TEMPLATE:  Added SSL connection entries.
 - config/search.py.TEMPLATE:  Added entry to connect to RabbitMQ cluster.
 - Documentation updates.